Feat: added delete option
Now with delete, the core functionality to the app is there.
It allows you to add in your own katas with tests, and have
the app pick you a kata to do.

Next steps are some polishing and adding a spaced repetition
algorith to hit harder things more frequently and also assign
a difficulty to the kata